requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
coveralls==3.0.1
cycler==0.10.0
docopt==0.6.2
idna==2.10
importlib-metadata==3.10.0
iniconfig==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.2
packaging==20.9
pandas==1.2.3
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
tabulate==0.8.9
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
